create CounterfactualChartWithLegend
